// src/redux/socials/socials.types.ts
export enum SocialsActionTypes {
  GET_FRIENDS_START = 'socials/GET_FRIENDS_START',
  GET_FRIENDS_SUCCESS = 'socials/GET_FRIENDS_SUCCESS',
  GET_FRIENDS_FAILURE = 'socials/GET_FRIENDS_FAILURE',
  GET_FRIEND_REQUESTS_START = 'socials/GET_FRIEND_REQUESTS_START',
  GET_FRIEND_REQUESTS_SUCCESS = 'socials/GET_FRIEND_REQUESTS_SUCCESS',
  GET_FRIEND_REQUESTS_FAILURE = 'socials/GET_FRIEND_REQUESTS_FAILURE',
  GET_SUBSCRIPTIONS_START = 'socials/GET_SUBSCRIPTIONS_START',
  GET_SUBSCRIPTIONS_SUCCESS = 'socials/GET_SUBSCRIPTIONS_SUCCESS',
  GET_SUBSCRIPTIONS_FAILURE = 'socials/GET_SUBSCRIPTIONS_FAILURE',
  SHOW_FRIENDS = 'socials/SHOW_FRIENDS',
  SHOW_FRIEND_REQUESTS = 'socials/SHOW_FRIEND_REQUESTS',
  SHOW_SUBSCRIPTIONS = 'socials/SHOW_SUBSCRIPTIONS',
  REMOVE_FRIEND_LOCAL = 'socials/REMOVE_FRIEND_LOCAL',
  REMOVE_FRIEND_START = 'socials/REMOVE_FRIEND_START',
  REMOVE_FRIEND_SUCCESS = 'socials/REMOVE_FRIEND_SUCCESS',
  REMOVE_FRIEND_FAILURE = 'socials/REMOVE_FRIEND_FAILURE',
  REJECT_FRIEND_REQUEST_LOCAL = 'socials/REJECT_FRIEND_REQUEST_LOCAL',
  REJECT_FRIEND_REQUEST_START = 'socials/REJECT_FRIEND_REQUEST_START',
  REJECT_FRIEND_REQUEST_SUCCESS = 'socials/REJECT_FRIEND_REQUEST_SUCCESS',
  REJECT_FRIEND_REQUEST_FAILURE = 'socials/REJECT_FRIEND_REQUEST_FAILURE',
  ACCEPT_FRIEND_REQUEST_LOCAL = 'socials/ACCEPT_FRIEND_REQUEST_LOCAL',
  ACCEPT_FRIEND_REQUEST_START = 'socials/ACCEPT_FRIEND_REQUEST_START',
  ACCEPT_FRIEND_REQUEST_SUCCESS = 'socials/ACCEPT_FRIEND_REQUEST_SUCCESS',
  ACCEPT_FRIEND_REQUEST_FAILURE = 'socials/ACCEPT_FRIEND_REQUEST_FAILURE',
  ADD_FRIEND_START = 'socials/ADD_FRIEND_START',
  ADD_FRIEND_SUCCESS = 'socials/ADD_FRIEND_SUCCESS',
  ADD_FRIEND_FAILURE = 'socials/ADD_FRIEND_FAILURE',
}